backend/requirements.txt
# ======== Core ==========
fastapi
uvicorn
httpx

sqlmodel
pydantic[email,timezone]
pydantic-settings
python-multipart


# ========== SDK ============
boto3


# ======== Security ==========
PyJWT
passlib[bcrypt]


# ======== AI ===========
# openai


# ======== Others ===========
pika
